+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory.swift
@@ -28,24 +28,29 @@ class PaymentSheetFormFactory {
     }
     let saveMode: SaveMode
     let paymentMethod: PaymentSheet.PaymentMethodType
-    let intent: Intent
-    let configuration: PaymentSheet.Configuration
+    let configuration: PaymentSheetFormFactoryConfig
     let addressSpecProvider: AddressSpecProvider
     let offerSaveToLinkWhenSupported: Bool
     let linkAccount: PaymentSheetLinkAccount?
     let previousCustomerInput: IntentConfirmParams?
 
+    let supportsLinkCard: Bool
+    let isPaymentIntent: Bool
+    let currency: String?
+    let amount: Int?
+    let countryCode: String?
+
     var canSaveToLink: Bool {
-        return (intent.supportsLinkCard && paymentMethod == .card && saveMode != .merchantRequired)
+        return (supportsLinkCard && paymentMethod == .card && saveMode != .merchantRequired)
     }
 
     var theme: ElementsUITheme {
         return configuration.appearance.asElementsTheme
     }
 
-    init(
+    convenience init(
         intent: Intent,
-        configuration: PaymentSheet.Configuration,
+        configuration: PaymentSheetFormFactoryConfig,
         paymentMethod: PaymentSheet.PaymentMethodType,
         previousCustomerInput: IntentConfirmParams? = nil,
         addressSpecProvider: AddressSpecProvider = .shared,
@@ -53,7 +58,7 @@ class PaymentSheetFormFactory {
         linkAccount: PaymentSheetLinkAccount? = nil
     ) {
         func saveModeFor(merchantRequiresSave: Bool) -> SaveMode {
-            let hasCustomer = configuration.customer != nil
+            let hasCustomer = configuration.hasCustomer
             let supportsSaveForFutureUseCheckbox = paymentMethod.supportsSaveForFutureUseCheckbox()
             switch (merchantRequiresSave, hasCustomer, supportsSaveForFutureUseCheckbox) {
             case (true, _, _):
@@ -66,7 +71,7 @@ class PaymentSheetFormFactory {
                 return .none
             }
         }
-
+        var saveMode: SaveMode
         switch intent {
         case let .paymentIntent(paymentIntent):
             saveMode = saveModeFor(merchantRequiresSave: paymentIntent.setupFutureUsage != .none)
@@ -79,9 +84,35 @@ class PaymentSheetFormFactory {
             case .setup:
                 saveMode = .merchantRequired
             }
-
         }
-        self.intent = intent
+        self.init(configuration: configuration,
+                  paymentMethod: paymentMethod,
+                  previousCustomerInput: previousCustomerInput,
+                  addressSpecProvider: addressSpecProvider,
+                  offerSaveToLinkWhenSupported: offerSaveToLinkWhenSupported,
+                  linkAccount: linkAccount,
+                  supportsLinkCard: intent.supportsLinkCard,
+                  isPaymentIntent: intent.isPaymentIntent,
+                  currency: intent.currency,
+                  amount: intent.amount,
+                  countryCode: intent.countryCode,
+                  saveMode: saveMode)
+    }
+
+    required init(
+        configuration: PaymentSheetFormFactoryConfig,
+        paymentMethod: PaymentSheet.PaymentMethodType,
+        previousCustomerInput: IntentConfirmParams? = nil,
+        addressSpecProvider: AddressSpecProvider = .shared,
+        offerSaveToLinkWhenSupported: Bool = false,
+        linkAccount: PaymentSheetLinkAccount? = nil,
+        supportsLinkCard: Bool,
+        isPaymentIntent: Bool,
+        currency: String?,
+        amount: Int?,
+        countryCode: String?,
+        saveMode: SaveMode
+    ) {
         self.configuration = configuration
         self.paymentMethod = paymentMethod
         self.addressSpecProvider = addressSpecProvider
@@ -93,6 +124,12 @@ class PaymentSheetFormFactory {
         } else {
             self.previousCustomerInput = nil
         }
+        self.supportsLinkCard = supportsLinkCard
+        self.isPaymentIntent = isPaymentIntent
+        self.currency = currency
+        self.amount = amount
+        self.countryCode = countryCode
+        self.saveMode = saveMode
     }
 
     func make() -> PaymentMethodElement {
@@ -113,7 +150,7 @@ class PaymentSheetFormFactory {
             additionalElements = [makeCashAppMandate()]
         } else if paymentMethod.stpPaymentMethodType == .payPal && saveMode == .merchantRequired {
             // Paypal requires mandate when setting up
-            additionalElements = [makePaypalMandate(intent: intent)]
+            additionalElements = [makePaypalMandate()]
         }
 
         // 2. Element-based forms defined in JSON
@@ -244,9 +281,9 @@ extension PaymentSheetFormFactory {
         return makeMandate(mandateText: mandateText)
     }
 
-    func makePaypalMandate(intent: Intent) -> PaymentMethodElement {
+    func makePaypalMandate() -> PaymentMethodElement {
         let mandateText: String = {
-            if intent.isPaymentIntent {
+            if isPaymentIntent {
                 return String(format: String.Localized.paypal_mandate_text_payment, configuration.merchantDisplayName)
             } else {
                 return String(format: String.Localized.paypal_mandate_text_setup, configuration.merchantDisplayName)
@@ -419,7 +456,7 @@ extension PaymentSheetFormFactory {
     }
 
     func makeAfterpayClearpayHeader() -> StaticElement? {
-        guard let amount = intent.amount, let currency = intent.currency else {
+        guard let amount = amount, let currency = currency else {
             assertionFailure("After requires a non-nil amount and currency")
             return nil
         }
@@ -434,7 +471,7 @@ extension PaymentSheetFormFactory {
     }
 
     func makeKlarnaCountry(apiPath: String? = nil) -> PaymentMethodElement? {
-        guard let currency = intent.currency else {
+        guard let currency = currency else {
             assertionFailure("Klarna requires a non-nil currency")
             return nil
         }